io.py: implemented nested list to numpy array conversion in val_unit_to_quantities 	   make strings be saved under 'val' key in quantities_to_val_unit        continue implementing loading rountine for h5 files
We forgot to replace some printing check in unit_to_quantities by its actual
functionality. Namely, replacing a nested list by a numpy array.

Additionally, strings should be represented under a 'val' key just
as we defined it in the yaml files.

Finally, we are trying to implement a test for whether the needed file is
already existing or not.