network.py: implement calculation of dependen network params
Here we implement the method that is called on initiation of the
Network class, that calculates all network parameters that are
derived from the ones specified in the yaml file.

Additionally we implemented a conversion of the weights to the ones
used in the theoretical papers and the implemented functions in
analytics etc.