meanfield_calcs.py: correct firing_rates, add doc, change transfer_function output
We had to correct a mistake in the calculation of the firing rates,
where mu and sigma need to be calculated each time of the recursion.
We added documentation, mainly for the transfer functions. And we
changed the output of the transfer function s.t. the conversion from
a nested list of quantities to a list of quantities of np.ndarrays is
done withing meanfield_calcs.