network.py: change decorator such that additional arguments are allowed
When the sensitivity measure is called, one needs to pass a
frequency for which the measure should be calculated. But the
corresponding result should be stored such that it is possible to
uniquely identify the frequencies that have been investigated.
For this reason, we now add those frequencies to the analysis
parameters.