network.py: new way of storing, saving and loading results
Networks now store all results in the results_hash_dict. The results
dictionary only contains the most recently calculated results, which
solves the problem of calculating a function with different methods.

The saving routine now saves the network_params, analysis_params,
results and results_hash_dict in an h5 file. This allows to completely
restore the network. The loading was adjusted appropriately.

The option of creating a Network object from a file was added.