Add new function for caching results
The decorator `_check_and_store` is problematic, because it does not
notice changes in the `analysis_params` dict. Therefore here we
introduce a new way of caching, which takes into account all
parameters passed to the cached function. This simplifies the caching
function and makes previously required arguments like `depends_on`
superfluous.