Fix input_output tests by adjusting cache
Currently `_cache` was converting the `result_keys` using
`np.atleast_1d`. This makes the strings numpy strings. When
saving a network using the `h5py_wrapper` the data types of all
keys are stored. Numpy strings are stored as `str_`. But when
loading the h5 file, the `h5py_wrapper` doesn't recognize this
data type because it only searches for `str` and `string_`.

Maybe numpy changed since the `h5py_wrapper` was written, but
this is a bug.