Introduce unit handling for results
The result units can now be added to the `_cache` function. They are
then stored in the new `result_units` dictionary of the corresponding
network. The order of arguments of `_cache` has been changed, because
`units` is an optional parameter and the function related arguments
should be grouped together. This required changes in many files.