Move delay_dist_matrix to `network_properties.py`
The delay distribution was moved to `models/utils.py` first, but
it seems unnatural. `models` should only contain the models themselves.
The delay distribution matrix is a network property that one might
need to recalculate once the analysis frequencies have been changed.
Therefore it should be in a place where calling it is more natural.
Hence we introduce `network_properties.py`.